api/requirements.txt
fastapi==0.68.1
pulsar-client==2.7.2
uvicorn==0.15.0